구입 후 처리 일부 구현
1. 구입 후 처리 일부 구현
- 미구입 시 턴 넘기기, 구입 시 돈 차감 및 땅 정보 업데이트 및 해당 정보 클라로 전송, 클라는 돈 텍스트 업데이트 까지 구현

-> 동기화, 구입 시 클라 업데이트(땅에 색칠), 건물 구입 미 구현

+++ BlueMarble_Server/MapManager.cpp
@@ -34,7 +34,7 @@ void MapManager::LoadMapData()
 
 	int code;
 	string name;
-	double landPrice = 0.0, villaPrice = 0.0, buildingPrice = 0.0, hotelPrice = 0.0, landMarkPrice = 0.0;
+	float landPrice = 0.0f, villaPrice = 0.0f, buildingPrice = 0.0f, hotelPrice = 0.0f, landMarkPrice = 0.0f;
 
 	readFile.open(mapFilePath);
 	if (readFile.is_open())
